reactor: allow more poll time in virtualized environments
In a virtual machine, IPIs can be significantly slower.  This slows down
both the signals used to wake sleeping shards, and the mprotect() calls
used to insert the memory barriers in non-sleeping shards.  This slowdown
dominates over the poll time, and the system ends up spending its time
in tgkill() and mprotect().

Extend the poll time in a virtualized environment to prevents this.