reactor: introduce scheduling_group and scheduled_function
A scheduling_group is a tag that can be used to mark functions; a
scheduled_function is such a marked function.  Each scheduling_group
can be assigned shares, and the scheduler will attempt to divide cpu
time between shares.

The scheduler will account functions according to the following table:

  marking        inline        group
  -------------------------------------------------
  unmarked       no            default group
  unmarked       yes           any group
  marked         dont care     owning group

This allows glue continuations that are executed inline not to cause
deferrals.

The scheduler instantiates one task queue per scheduling group, and
alternates between queues according to share-weighted runtime, subject
to exponentuial decay.

scheduled_function is implemented as a thunk; executing it while in the
wrong queue will cause it to defer.  It is intended that primitives
such as repeat_until_value() will have specializations for scheduled_function,
allowing the thunking to be avoided.