Optimize exception_ptr
std::exception_ptr's destructor is out-of-line, which reduces inlining
opportunities when passing futures around, as each such operation must
be prepared for the future to be exceptional and therefore call the
exception_ptr destructor.

Fix this by introducing a wrapper around std::exception_ptr, called
seastar::exception_ptr, which delays the actual destruction be queueing
exception_ptrs that are to be destroyed.  The space for the queue is
pre-allocated when the exception_ptr is created (a rare event) and so
only three instructions are needed for the destructor.

Testing with scylla's perf_simple_query shows mixed results.  Dynamic
instruction counts are down 3%, but performance is also down 1%; this needs
to be investigated later not on a laptop.

Also TODO: documentation.