cpu scheduler: correctly order requests
The comparison function is wrong, and we are ordering the requests from
larger vruntime to smaller. That can be easily seen by adding some debug
code that prints the state of the array before and after sorting.

Before the patch:

	I have active tasks, and they are
		 name main, vruntime 31216
		 name sg50, vruntime 0
		 name sg20, vruntime 0
		 name sg100, vruntime 0
	After partial sort: active tasks, and they are
		 name main, vruntime 31216
		 name sg50, vruntime 0
		 name sg100, vruntime 0
		 name sg20, vruntime 0

After the patch:

	planning a run, ticks 8 active queues 4
	I have active tasks, and they are
		 name main, vruntime 31185
		 name sg50, vruntime 0
		 name sg20, vruntime 0
		 name sg100, vruntime 0
	After partial sort: active tasks, and they are
		 name sg50, vruntime 0
		 name sg100, vruntime 0
		 name sg20, vruntime 0
		 name main, vruntime 31185

Note how now one of the tasks with smaller vruntime is in the
front, whereas before main, with the largest vruntime was kept
in the front.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <1492824785-25334-1-git-send-email-glauber@scylladb.com>