cpu scheduler: make _last_vruntime monotonic
_last_vruntime is used to track what was the largest vruntime we have
seen, so that fair sleepers will catch up with that.  It works, because
we will order classes by ascending estimated vruntime, so a class that
runs after another class should always have a larger vruntime by the end
of the run.

However, things not always go according to the plan. For instance, a
class that is full of very light tasks could be planned to run for 2 or
3 ticks, but in reality run for less than even one. If that is the last
class in the plan to run, _last_vruntime can go back by a lot, since the
previous classes will have advanced it considerably already - specially
if they are heavy.

A task being inserted at this point will gain advantage over the other
heavy classes that ran earlier.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <1493094608-3967-1-git-send-email-glauber@scylladb.com>