future-util: add scheduled_function variants of repeat() and friends
seastar::scheduled_function<> was designed to support being called as an
rvalue (usual temporary passed to another function) and as an lvalue (with
the expectation that the caller has taken care of persisting it, and with
the ability to call it multiple times).  However, this fails in repeat()
when called with an rvalue.  repeat() will need to access its input
AsyncAction twice: once to call it for an iteration, and another to pass it
along for the next iteration.  Clearly, they can't both be called as rvalues,
so the first call, as an lvalue, ends up accessing a dangling reference.

While this represents a failure of the scheduled_function model, fixing it
is deferred for now.  Instead, we provide overloads of the repeat() family
for scheduled_function that unwrap and re-wrap it correctly.