reactor: introduce scheduling
Instead of having a single queue of tasks to run, allow
users to create multiple task queues, each associated with
a `scheduling_group` and a weight. The scheduler will try to run
tasks from different queues so that the overall runtime is
distributed according to the assigned weight.

Currently, the only way to tag a task is with make_task(), but we
will add more methods in future patches.