task: resumable tasks
Currently, a task object is single use. After calling task::run(), the
reactor deletes the task, so it cannot be re-entered.

This patch lays the groundwork to re-entering tasks that were already
run once by moving the deletion of the task from the reactor to task::run()
itself.  Normally, a task will delete itself after running, but it may
also choose not to do so, and it can then schedule itself again.

To avoid confusion, task::run() is renamed to task::run_and_dispose().

Good candidates for resumable tasks are looping primitives like repeat()
and C++ coroutines.