future: make class continuation resumable
Add support for resuming a continuation by splitting it into
a base class continuation_base (that adds a future_state to class task)
and continuation (that adds the lambda that processes the state to
class continuation_base).

Enhance future with a set_callback(continuation_base) method that allows
chaining a future into a resumable task.