future-util: deprecate primitives that forward functions as references
Forwarding a function across a continuation that was passed via a
reference is dangerous, as it can easily lead to use-after-free if
inadvertant. To prevent this, create overloads that accept references
and deprecate them. Callers made arrangements for the functions to be
stable across continuations can indicate this by using std::ref().

Primitives that do not pass functions across continuations (like
parallel_for_each) are not updated.