systemwide_memory_barrier: add and use try_systemwide_memory_barrier
The current implementation of systemwide_memory_barrier causes
contention in the kernel, so we protect it with a trylock. However,
a new implementation (using the membarrier() system call) does not
cause contention, so the trylock becomes unnecessary (and a cause of
contention in itself).

To allow adding a new membarrier implementation, move the trylock
into the systemwide_memory_barrier API. The API can then elide the
lock if it is unneeded.