linux-aio: try to perform io_getevents in userspace
If enough events are available (or if the timeout is zero), we
can perform io_getevents entirely in userspace. This is possible
because the completion event ring is mapped into userspace using
a stable data structure.

libaio implements this for the special case of zero available events.

Credit to QEMU for pioneering this optimization.