reactor: prepare preemption mechanism for alternative signalling
In essence, our preemption mechanism exposes a periodic kernel
timer to the reactor in a way that can be efficiently polled. We
started with a signal whose handler set a thread-local variable (but
that did not work too well due the non-scalability of Linux signals).
We switched to a side-thread that sleeps on a timer and then writes
to that memory location, but that causes many context switches.

This patch prepares the way for a third method. We note that
the aio completion ring (which is read by io_getevents()) is written
to by the kernel, and a future patch will set things up so that the
ring will be written to by our periodic timer.  To prepare for this,
we lay out the thread-local memory location in the same format as
the aio ring (a pair of 32-bit indexes, with the convention that when
they are unequal, we need to break out of the task execution loop).

For now (and for the foreseeable future) we will still support the
side thread. Having a compatible memory layout means we can introduce
aio polling as an alternative, without requiring virtual functions in
the fast path.