reactor: switch smp notifications from signals to eventfds
While signals are generally nasty, in this case the motivation to
remove them is to allow the main event loop to use io_getevents().
Since io_getevents() has no signal-enabling equivalent (like
pselect() or epoll_pwait()), we can't use signals reliably for
waking up.

Switch to eventfd instead, which should also be faster. File
descriptors should not be used with dpdk, but since with dpdk
we never sleep, we also never need to wake up.