reactor: switch smp notifications from signals to eventfds
Signals are generally nasty, requiring a process-wide lock to send.
This makes them slow on large machines (although this is mitigated
by the reactor spinning before going to sleep).

Switch to eventfd instead, which should also be faster. File
descriptors should not be used with dpdk, but since with dpdk
we never sleep, we also never need to wake up.

The original motivation for this patch was the lack of io_pgetevents(),
which is needed to atomically wait for both signals and fd
notifications. Since io_pgetevents() was since made available, this
patch is not critical, but is still an improvement.