reactor: virtualize high resolution timer processing
The epoll backend uses signals for high resolution timers, so that a timer
expiration will set need_preempt(). But since we'll use a different method for
preemption, we need to move away from signals with the aio backend.

To prepare for that, move high resolution timer processing to reactor_backend.