reactor: add sampler for reactor stalls
Use the getrusage() system call to sample the number of voluntary
context switches, which should be zero if there are no stalls.
Exclude reactor sleeps, which are non-stalling context switches
(there is nothing to do, so we yield the cpu).

This could be dynamically enabled in the future and exported via
metrics, so we can find out if there are stalls dynamically.