linux-aio: stop including <linux/aio_abi.h>
In Linux 4.18, the linux/aio_abi.h header conflicts with
glibc headers (different, though of course compatible, definitions
of struct timespec).  Use our own structures instead; since it's
an ABI, as long as the layout is compatible it should keep working.

To reduce the chance of future conflicts, place ABI names in a
linux_abi namespace so if some random #include brings the originals
into scope, we won't conflict.