future: don't check for preemption in future::then() and future::then_wrapped()
There are two typical .then() continuations styles: one is supposed to process
the results of the previous promise and launch a new one - i.e. it is a linear
process without significant loops, so it takes a small amount of time compared
to the task quota. The other style contains loops and is therefore not bounded in
compute time.

If we assume the first style indeed takes a trivial amount of time, and the second
uses Seastar primitives such as repeat() and do_for_each(), then it follows that
we can remove the preemption check in future::then() and future::then_wrapped().
A continuation chain will conform to one of the following:

 - it is composed of trivial computations that together take a trivial amount
   of time (compared to the task quota)
 - one of the continuations defers (and therefore preemption is not needed)
 - one of the continuations contains a Seastar looping primitive, and will
   therefore preempt if needed.

An exception to this idealic description is loops performed using tail recursion
rather than Seastar primitives. These loops will now be translated into a long
chain of trivial computations, and will so be able to exceed the task quota without
being preempted, causing latency spikes. Since removing the calls to need_preempt()
significantly reduce the dynamic instruction count, we will treat such loops as
bugs and convert them to Seastar looping primitives instead.

We still preempt in debug mode, in order to catch lifetime problems that won't
be caught when the future used is nearly always ready.

The test case test_high_priority_task_runs_before_ready_continuations will fail
after this, so it is removed. Its users do not rely on such tasks being scheduled
ahead of .then() continuations.