rpc: connection-level resource isolation
Currently, rpc allows a scheduling_group to be configured on a per-verb basis.
This method has some drawbacks:
 - if different verbs in the same connection are executed under different
   groups, head-of-line blocking can cause a low-priority group to block a
   high priority group by consuming all of the units in _resources_availble
 - the same verb can be executed under only one group, making multi-tenancy
   harder.

This patch adds a new method, which isolates a entire connection rather than
a verb. The new method improves upon the old method by:
 - allowing the same verb to be executed under different groups (if called
   under different connections)
 - executing more server code under scheduling groups
 - preventing head-of-line blocking (requires more work to have per-group
   memory reservations)

Eventually, we'll deprecate and remove the old method.