sharded: type-erase invoke_on_all()
sharded<>::invoke_on_all() is used for the slow path: it is used for
replicating metadata or for global control changes rather than the
data path. As such it can be unconditionally type-erased to reduce
compile time. This is especially important for invoke_on_all as its input
function gets compiled many times as it is propagated down the call chain.