reactor: add command-line flag to force aio syscalls
Seastar attempts to bypass aio syscalls when possible, taking advantage of the
event ring. However, that makes debugging with strace harder as it is impossible
to see when and how the kernel responded to io_submit() calls.

Improve that by adding an option to force aio syscalls to enter the kernel.