tests: dns: protect against home routers that grab and respond to all dns requests
My home router responds to dns requests to 1.2.3.4:53, even though there is
no such dns server; in fact it ignores the destination IP address completely.
This makes the timeout test, that expects resolution to fail, to get a response
and report a failure.

Work around this by setting a fake udp port in addition to the fake address.