future: optionally type-erase future::then() and future::then_wrapped
The future::then() and future::then_wrapped() templates are usually called
with a distinct Func type per invocation (since these types are usually
lambdas). That means that the code in then() and then_wrapped() is compiled
as different functions for each invocation. Moreover, Func::operator() is
invoked twice, so if the compiler starts inlining, even more code is compiled.
The result is an explosion of generated code and compile time.

To combat this, provide a mode where Func is type-erased before the meat
of then() and then_wrapped() is called. That reduces the amount of generated
code significantly. We also call it just once to discourage the compiler
from attempting to inline in. This results in a 37% decrease in generated
code (and presumably compile time) on a large source file.

Note that type erasure is partial: we still have a version of then() for
each future specialiation, and for each result type. But there's far fewer
of these combinations that individual invocations of these functions.