sharded: support for non-cooperative service types
The sharded<Service> template was designed with cooperating service classes
in mind. This cooperation is in the form of providing a stop() method for
orderly shutdown.

The sharded<> template is also useful for non-cooperating services: for example,
std::unordered_map<std::string can be broadcast to all shards with

    sharded<std::unordered_map<sstring, sstring>> my_map;
    my_map.start(initializer).get();
    ...
    my_map.invoke_on_all([] (my_map_type& mm) { mm["x"] = "y"; }).get();
    ...
    my_map.stop().get();

All is required is that the "Service" doesn't perform asynchronous operations (as
is usual for plain data types) or that the user stops them in some other way. But
this currently doesn't compile because the Service::stop() method is required.

To allow for with, we provide a sharded_adapter<> template. If the user wraps their
service with sharded_adapter, then stop() is no longer needed and you can distribute
any type.