memory: add missing attributes to malloc famility function aliases
gcc9 now demands that function attributes on aliases match the implicit
function attributes on their target. Humor it.

Note that the [[gnu::leaf]] attribute is actually a lie - we do call back
from malloc(), via the reclaim mechanism. But reclaim already has to promise
not to affect whoever called malloc (which happens in Scylla by preventing
reclaim while touching LSA memory). In any case the attributes in memory.cc
have no affect on this; rather the attributes on the declration seen
at the call site and implied matter.