core: add helper for parsing and comparing kernel versions
In the past we tried to detect kernel features by discovering whether the feature
exists and works properly. But this is quite hard for some features, because the
error codes returned when the feature does not exist are not unique (EINVAL from
some aio features can also be returned when another parameter is invalid) or
because the feature requires writing, and we don't have a path we can modify. In
addition, the strategy fails for bugs that cannot be triggered reliably.

So instead of detecting features, we'll just parse and compare kernel versions.
This is elaborate because kernel history is not linear: bug fixes can be backported
to older versions. For example, a fix that landed in 4.12 may be backported into
4.10.20 (but not 4.11.anything or 4.10.19), and may be backported into
distribution-specific kernels as well.

The way to approach this is to divide the version string into a sequence of
components, and require that all components until the last be equal (these are
used to identify the branch) and the final component be greater than or equal to
(this is used to identify the position in that branch's history). The first two
components are treated specially since they are linear (5.1 contains all of 4.19).