future-util: reduce allocations and continuations in parallel_for_each
Currently, parallel_for_each requires a continuation for every non-ready future
it processes. We can reduce this with resumable tasks by allocating just one task
for all non-ready continuations and reusing it for each non-ready future.
Furthermore, when a future is completed, other futures may have also been completed,
so there is no need to schedule the task for them. This was, the number of tasks
allocated is reduced from N to 1 and the number of tasks executed can be anywhere in
the range [1, N] depending on run-time characteristics.