reactor: replace task queue circular buffer with a linked list of small arrays
Currently, we use a circular_buffer to store tasks awaiting
execution. This is efficient because we amortize many task pointers
over a cache line's worth, and the CPU can easily prefetch ahead.

However, a circular_buffer only allows queueing tasks at the
front and back of the queue, while we wish to sort tasks according to
their type. As a first step, replace the circular buffer with a linked
list of arrays (each occupying two cache lines). Each queue fragment
is relatively efficient, and in addition to queueing at the front and
back, we can queue to any non-full queue fragment (which we'll start
doing in the next patch).