temporary_buffer: add support for coalescing
Introduce temporary_buffer::try_coalesce_with(second_buffer), which
attempts to merge two buffers that are derivatives (in terms of
temporary_buffer::share() and similar) of the same original buffer
and which are contiguous in memory.

The motivation is to help output_stream::flush(), which is forced
to split a buffer at the flush point, to be able to unsplit the
buffer if is has the opportunity to do so. Currently flush()
does this by delaying the splitting until the next poll period,
but we wish to change this (details to be explained in the patch
that modifies output_stream).

To support coalescing, we must check whether the two buffers
originally came from the same allocation (it is not enough
to check if they are adjacent in memory as that can happen
accidentally). For that we add a new member _offset_in_orig_buffer.
We lose the ability to have 4 GiB temporary_buffers, but that's
an abuse of temporary_buffers anyway.