reactor: add an internal facility for scheduling tasks after a poll period
In order to batch work, we want to delay it until the next poll period. This
adds a facility for doing so.

It is similar to execution_stages, but lower-level and more general:
 - it is not restricted to calling functions that return a future
 - it works directly with tasks

Scheduling tasks is better than running directly in a poller, since
the task runtime is accounted for, and tasks are executed in the
correct scheduling_group. If a scheduling_group has low priority,
the task can be scheduled long after the poller executed, increasing
batching and scheduling fidelity.

We can probably update execution_stages to work with this.

Note that while this is adding a poller, the series will remove a poller
later on, preserving the poller balance.