pollable_fd: expose no_more_recv/no_more_send
Later patches will increase batching of writes, so we may end up not calling
the send() system call immediately. This defeats connect_test.cc's
test_unconnected_socket_shutsdown_established_connection, which
uses a buffer size of 1 to force the system call to be issued immediately.

To preserve the current behavior, we'll expose no_more_send so that the check
can be performed earlier, when we queue the write, rather than later, when
we actually send the syscall, too late for the test.