future, future-utils: stop returning a variadic future from when_all_succeed
when_all_succeed() can return a variadic future (if two or more
of its input futures are non-void). This is a problem since
variadic futures are deprecated and to be removed.

This patch introduces a temporary solution:
 1. if when_all_succeed() returned a nullary or unary future, nothing is
    changed.
 2. otherwise, instead of returning a variadic future, it returns a
    future<internal::when_all_succeed_tuple<T...>> instead of future<T...>.
 3. The newly-introduced internal::when_all_succeed_tuple<> template is
    just a thin wrapper around std::tuple<>. The reason we need a wrapper
    is that so then() knows to treat it specially.
 4. If future::then() is called on a future carrying a
    when_all_succeed_tuple, then it unpacks the tuple into the lambda's
    parameters (same as the behavior for the variadic future it replaces).

Returning an internal type from when_all_succeed() is gross, but I expect
all of the usages of when_all_succeed() call .then() immediately, since
when_all_succeed() is used to collect the results of several fibers.
There isn't much else you can do with it. And in fact Scylla compiles with
this patch (both dev and release modes tested).

I plan to follow up with an API change to make when_all_succeed() return
an ordinary std::tuple<>, at the user's option.