future-util: make when_all_succeed() return a future<std::tuple<...>>
Originally, when_all_succeed() returned a variadic future. Variadic
futures are now deprecated, so support for returning a tuple-like
type when_all_succeed_tuple was added, when the number of values
to be returned was greater than 1. In addition, the previous
behavior of unpacking the tuple to the continuation parameter list
in future::then() was emulated.

This behavior is too magical:
 - the type is a tuple or not a tuple depending on the arity
 - the type-dependent unpacking is irregular

To reduce irregularity, this patch introduces a new api level with
more regular semantics: when_all_succeed() always returns a tuple.
The tuple be unpacked with future::then_unpacked(), without requiring
the magic when_unpacked().

The types used for the previous API level are deprecated.