coroutine: deprecated forwarding another function's returned future
The syntax

    co_return a_function_that_returns_a_future(p1, p2);

saves a task switch, but is a time bomb as any parameters (p1, p2)
are not held stable on the coroutine frame and instead must be copied
or moved. Since it's easy to invoke this accidentally, and since the
syntax is not legal for the future<void> case, deprecate it.

There is no replacement yet, apart from the less-efficient
"co_return co_await ...".