cmake: tighten compiler flag checks for clang++
Clang++ reports a warning for a compiler flag that it recognizes but
cannot use. An example is -fstack-clash-protection that is recognized,
but not implemented on aarch64 (it is implemented on x86_64). The
compilation still terminates with an error code of 0.

Cmake has a list of error patterns to check for, but that list
does not include the warning that clang produces, so cmake thinks
that clang accepted the flag. This later fails when it is used
together with -Werror.

Fix by adding that pattern to the compiler invocation.

I filed [1] so that future cmake versions are fixed.

[1] https://gitlab.kitware.com/cmake/cmake/-/issues/21998