smp: allow having multiple instances of the smp class
Seastar is often used for clustered systems, so it is
desirable to run multiple Seastar instances in the same
process for testing, to simulate a cluster.

As a first step in that direction, the smp class made
free of globals, so we can create multiple instances.
In most cases this means just removing "static" from
the variable declaration, but there are some exceptions:

1. smp::_qs is used for the static interface
   smp::invoke_on(), so it is made thread-local (with
   every thread that hosts a reactor belonging to the
   same smp instance having a copy of the variable).
   Since _qs has a destructor that needs to be run only
   once, the destructor is moved to a new variable
   smp::_qs_owner that is a regular data member.

2. smp::count is a public interface, and cannot be
   made thread_local since it is used by alien_test
   outside reactor threads (and presumably, possibly
   by applications that use seastar::alien). So it is
   kept global. The implication is that programs with
   multiple smp instances must have the same smp::count
   across all instances, until this interface is
   deprecated and removed.

3. smp::_tmain is made thread_local as the simplest
   translation I could find. I plan to remove it later
   as it's used in only one unimportant place.

The smp instance is created by app_template. It is a
std::shared_ptr, since reactors need to hold on to it,
and reactor destruction happens at thread termination. This
may be simplified after we improve the shutdown sequence.

Note that this is only a first step, as globals must be
removed from other parts of Seastar (for example,
seastar::alien), but it is the most complicated step.

Tested in release and debug mode, with and without dpdk.

Credit to Konstantin Osipov <kostja@scylladb.com> for the
idea.