reactor: fault in memory during program startup if --lock-memory was specified
In c15055cbb8fe50 ("memory: add option to mlock() all memory"), we call
mlockall() in order to prevent paging from causing stalls while the
reactor waits to fault in memory. In 4d61dba522 ("reactor: don't mlock
all memory at once") we made this locking happen on first access,
to prevent huge startup delays on large machines.

However, this still leaves a window between process startup and until
all memory has been touched when faults can happen. The problem is
exacerbated by the kernel serializing faults on mmap_sem, and by using
transparent huge pages - if 16 shards touch new memory at the same time,
one shard will have to wait while 16*2MB=32MB of memory are being cleared.
This manifests in reactor stalls in semi-random places, see [1] for an
example.

To fix this, fault in pages eagerly. There is still a window of time
where we are faulting in pages, but that window is much narrower, and
hopefully overlaps the system initializing itself and joining a cluster.

To perform this more-eager initialization, we use a poller. Pollers
have the advantage of being called regularly, and being able to
utilize idle time. Each shard faults in its pages independently, but
first checks that other shards aren't faulting, to prevent serialization
on mmap_sem. The poller unregisteres itself if the reactor shuts down
or if it completes.

The whole thing is disabled if the Seastar allocator isn't used, or
it --lock-memory is not requested.

[1] https://github.com/scylladb/scylla/issues/8828