coroutines: introduce all() template for waiting for multiple futures
This patch introduces an all() function, similar to the
exeisting when_all_succeed() but designed for coroutines. It
also introduces a new #include directory (seastar/coroutine) and
a new namespace (seastar::coroutines), for a wider coroutine support
library.

Compared to when_all_succeed(), all() is faster - it does not have
an extra allocation for a state object, instead utilizing a coroutine's
ability to store temporary objects in the coroutine promise object
(which is allocated anyway). In terms of interface, it drops
when_all_succeed()'s ability to accept future objects (this is
very error prone, since if the coroutine throws before it reaches
all(), those futures will not be waited on). Instead, all() requires
all futures to be produced by functions that it invokes.

Unit tests are added.