circleci: relax docker limits to allow running with new toolchain
Switching to a new toolchain container tickles docker in a slightly
different way, so now it refuses to run with a message like

    Run Build Command(s):/usr/bin/ninja cmTC_32c16 && ninja: fatal: posix_spawn: Operation not permitted

This comes from docker attempting to confine the contained processes,
though why it would confine fork() or clone() is not clear to me.

Fix by removing seccomp confinement, we fully trust our own code.
Also add --pids-limit -1 for good measure; this was observed to prevent
fork() problems on Scylla's dbuild.