utils: log: don't crash on allocation failure while extending log buffer
We use an auto-extending log buffer, and the buffer extension can
fail if we run out of memory. This would be caught, except that
it can happen within some destructor in fmt, which in turn causes
std::terminate() to be called. This is bad.

Instead of crashing, catch the exception internally and brand the
message in a way that shows it is truncated. The relevant functions
are marked noexcept. To avoid more complexity about deciding whether
to append the input character or not, move the appending to the
function that reallocates the buffer, which already knows if we
can append or not.

See scylladb/scylla#10613.